Rename top-level component directory to GradientButton
